general refactor & move cmd type to scaffolder
