+++ starport/interface/cli/starport/cmd/app.go
@@ -23,8 +23,8 @@ func NewApp() *cobra.Command {
 func appHandler(cmd *cobra.Command, args []string) error {
 	name := args[0]
 	addressPrefix, _ := cmd.Flags().GetString("address-prefix")
-	sc := scaffolder.New(name)
-	path, err := sc.Init(scaffolder.AddressPrefix(addressPrefix))
+	sc := scaffolder.New(appPath)
+	path, err := sc.Init(name, scaffolder.AddressPrefix(addressPrefix))
 	if err != nil {
 		return err
 	}
