+++ starport/interface/cli/starport/cmd/type.go
@@ -1,19 +1,10 @@
 package starportcmd
 
 import (
-	"context"
 	"fmt"
-	"go/ast"
-	"go/parser"
-	"go/token"
-	"os"
-	"path/filepath"
-	"strings"
 
-	"github.com/gobuffalo/genny"
 	"github.com/spf13/cobra"
-	"github.com/tendermint/starport/starport/pkg/gomodulepath"
-	"github.com/tendermint/starport/starport/templates/typed"
+	"github.com/tendermint/starport/starport/services/scaffolder"
 )
 
 func NewType() *cobra.Command {
@@ -28,77 +19,10 @@ func NewType() *cobra.Command {
 }
 
 func typeHandler(cmd *cobra.Command, args []string) error {
-	path, err := gomodulepath.Parse(getModule(appPath))
-	if err != nil {
+	sc := scaffolder.New(appPath)
+	if err := sc.AddType(args[0], args[1:]...); err != nil {
 		return err
 	}
-	typeName := args[0]
-	ok, err := isTypeCreated(appPath, path.Package, typeName)
-	if err != nil {
-		return err
-	}
-	if ok {
-		return fmt.Errorf("%s type is already added.", typeName)
-	}
-	var fields []typed.Field
-	for _, f := range args[1:] {
-		fs := strings.Split(f, ":")
-		name := fs[0]
-		var datatype string
-		acceptedTypes := map[string]bool{
-			"string": true,
-			"bool":   true,
-			"int":    true,
-			"float":  true,
-		}
-		if len(fs) == 2 && acceptedTypes[fs[1]] {
-			datatype = fs[1]
-		} else {
-			datatype = "string"
-		}
-		field := typed.Field{Name: name, Datatype: datatype}
-		fields = append(fields, field)
-	}
-	g, _ := typed.New(&typed.Options{
-		ModulePath: path.RawPath,
-		AppName:    path.Package,
-		TypeName:   typeName,
-		Fields:     fields,
-	})
-	run := genny.WetRunner(context.Background())
-	run.With(g)
-	run.Run()
 	fmt.Printf("\n🎉 Created a type `%[1]v`.\n\n", args[0])
 	return nil
-}
-
-func isTypeCreated(appPath, appName, typeName string) (isCreated bool, err error) {
-	abspath, err := filepath.Abs(filepath.Join(appPath, "x", appName, "types"))
-	if err != nil {
-		return false, err
-	}
-	fset := token.NewFileSet()
-	all, err := parser.ParseDir(fset, abspath, func(os.FileInfo) bool { return true }, parser.ParseComments)
-	if err != nil {
-		return false, err
-	}
-	for _, pkg := range all {
-		for _, f := range pkg.Files {
-			ast.Inspect(f, func(x ast.Node) bool {
-				typeSpec, ok := x.(*ast.TypeSpec)
-				if !ok {
-					return true
-				}
-				if _, ok := typeSpec.Type.(*ast.StructType); !ok {
-					return true
-				}
-				if strings.Title(typeName) != typeSpec.Name.Name {
-					return true
-				}
-				isCreated = true
-				return false
-			})
-		}
-	}
-	return
 }
